Add files via upload
This program is used for a large demand of products to be shipped all over the country. The API is acting very slow and we must use concurrency so that the program can be amped up and be efficient and faster. I'm also calculating how much money is being spent within the shipping. Program has 2 files provided: ShippingLabel.java and ShippingService.java provided by instructor.